Fix cross-platform build for core (#69)
Introducing the DeepFilter linux .so as a dependency had broken the cross platform build of the MediaProcessor.
This fixes the issue for Windows and arm based macOS platforms.
There is still an issue with the x86_64 macOS that wouldn't work even with rosetta on arm. I'll have to check that further, but no reason to block this as it's an urgent fix.