Add audio support downstream (#76)
This PR extends support for audio files to downstream of MediaProcessor, i.e. the backend and frontend.

A media file type determination -with a bit of redundancy between the core- has been added to the backend.
This is mainly a fast fix that for something that was otherwise blocking direct audio support on the web application.
The functionality exists within the core, but not yet decided how it will be exposed.

In addition, the frontend was updated to differentiate between the two media types and represent them appropriately.