Refactor: move `splitAudioIntoChunks` as `splitMedia`
This commit moves the `splitAudioIntoChunks()` method to from `AudioController`
to `FFmpegController` as `splitMedia()`. The `populateChunkDurations` helper
was also moved to this class as a private member, while
`generateChunkFiles` (previously part of this class) is now also private. These
changes were necessary in order to 1) ensure `FFmpegController` has access to
the proper methods for testing, and 2) reduce friction for the user, as they
now only need to call `splitAudioIntoChunks` from this public API. We should
consider removing these methods from `AudioController` if the changes are
acceptable.