Improvement/optimize core runtime performance (#79)
This PR improves the runtime performance of the MediaProcessor by ~15% !

Profiling shows we're spending ~70% of our time filtering the data. Without touching the rust library internals, introduced optimizations on the compilation and that alone brought about an additional 10% reduction.
Adding similar optimizations to the core, as well as slightly better arranging data and resource usage in AudioProcessor, we've totalled out at about 15% faster processing.