Fix broken core on darwin (#85)
This PR addresses compatibility issues on Apple Silicon macOS machines when using Docker to emulate x86_64. The issue arose due to the libdf.so being compiled with AVX2 instructions for x86_64, which are apparently not supported by Docker's QEMU-based emulation on these systems. See the related discussion here for details.
Unfortunately, the QEMU limitation is still here, and removing this optimization from the library fixed this issue for us.

Additionally few minor updates to core, to fix a deprecation error with clang and a few other warnings.