Feature: Add ffmpegcontroller to core (#60) (#63)
This PR introduces an `FFmpegController` along with its test suite.
It also updates the `FFmpegCommandBuilder` and `FFmpegConfigManager` to facilitate the changes required by the controller. 
This implementation of the controller works closely with the `FFmpegConfigManager` and further relies on it for the validity of FFmpeg configuration fields. 
The config manager should be updated before merging this into main.